refactor: 타입 import 경로 수정 / DP-274
